Add files via upload
This code includes Tim's code. To test our solution compared to theirs, for now, run -- Controller InputFiles/<InputFile> OutputFiles/<OutputFile>. This will create and open their output file and our result shows in terminal and can then be compared. When we have created the writing to a file for our code, we can change the code in Controller to take 3 command line arguments and then open both outputfiles or even write code to compare them automatically. 

Right now, Branch and Bound seems to be working fine for the assignments. We are missing error messages/ handling. 

Note: sometimes our solution does not match their solution but the total cost is the same- thus both solutions are correct and valid.